STYLE: Replace unsigned long w/ unsigned int in Transformix GTest checks
Reduced the use of `unsigned long`, as it appears platform specific.

+++ Core/Main/GTesting/itkTransformixFilterGTest.cxx
@@ -1577,7 +1577,7 @@ GTEST_TEST(itkTransformixFilter, CheckMinimumMovingImageUsingAnyInternalPixelTyp
   };
 
   check(TypeHolder<char>{});
-  check(TypeHolder<unsigned long>{});
+  check(TypeHolder<unsigned int>{});
   check(TypeHolder<double>{});
 }
 
@@ -1619,7 +1619,7 @@ GTEST_TEST(itkTransformixFilter, CheckZeroFilledMovingImageWithRandomDomainUsing
   };
 
   check(TypeHolder<char>{});
-  check(TypeHolder<unsigned long>{});
+  check(TypeHolder<unsigned int>{});
   check(TypeHolder<double>{});
 }
